Refactoring + cleanup
