+++ src/store/hooks/index.ts
@@ -1,6 +1,7 @@
 export * from './useAnnotations';
 export * from './useDataModel';
 export * from './useExcelAnnotationExport';
+export * from './useExcelRelationshipExport';
 export * from './useFolderMetadata';
 export * from './useImageMetadata';
 export * from './useImages';
